feat(vulkan): implement VRAM reservation and fix warnings
- Added tracking to VulkanBackend with std::map and std::mutex

- Implemented set_reservation() to cap VRAM usage

- Updated create_buffer to respect reservation

- Updated VulkanCompute and VulkanOp to use tracked free_memory()

- Fixed compilation warnings in scheduler.h and ui.h

- Suppressed executable stack warning in CMakeLists.txt

+++ mm-rec-cpp-eigen/include/mm_rec/utils/ui.h
@@ -172,7 +172,7 @@ inline void print_header(const std::string& title, int width = 60) {
     std::cout << box::VERTICAL;
     for (int i = 0; i < padding; ++i) std::cout << " ";
     std::cout << color::BOLD << color::WHITE << title << color::CYAN;
-    for (int i = 0; i < width - title.length() - padding - 2; ++i) std::cout << " ";
+    for (int i = 0; i < width - static_cast<int>(title.length()) - padding - 2; ++i) std::cout << " ";
     std::cout << box::VERTICAL << "\n";
     
     std::cout << box::BOTTOM_LEFT;
